+++ applications/bot/src/commands/index.ts
@@ -5,6 +5,7 @@ import { FinanceCommand } from '~/commands/finance/index.js';
 import { InfoCommand } from '~/commands/info/index.js';
 import { MakanickoCommand } from '~/commands/makanicko/index.js';
 import { SummonCommand } from '~/commands/summon/index.js';
+import { TrainingCommand } from '~/commands/training/index.js';
 
 export const commandBuilder = Ix.builder
   .add(EventCommand)
@@ -12,4 +13,5 @@ export const commandBuilder = Ix.builder
   .add(FinanceCommand)
   .add(InfoCommand)
   .add(SummonCommand)
-  .add(CarpoolCommand);
+  .add(CarpoolCommand)
+  .add(TrainingCommand);